tests: fix error handling in test runner
Currently, the test runner can use variables in the `except` branch
before they're initialized. Hoist all the local state variables to
before the try: to fix that.

Also, make BEFORE_PID logic run only if the test uses `{{BEFORE_PID}}`.

+++ tests/runtime/engine/runner.py
@@ -158,10 +158,11 @@ class Runner(object):
 
         signal.signal(signal.SIGALRM, Runner.__handler)
 
+        p = None
+        before = None
+        bpftrace = None
+        after = None
         try:
-            before = None
-            bpftrace = None
-            after = None
             timeout = False
 
             print(ok("[ RUN      ] ") + "%s.%s" % (test.suite, test.name))
@@ -210,7 +211,7 @@ class Runner(object):
                             raise TimeoutError('Timed out waiting for BEFORE %s ', test.before)
 
             bpf_call = Runner.prepare_bpf_call(test)
-            if test.before:
+            if test.before and '{{BEFORE_PID}}' in bpf_call:
                 childpid = subprocess.Popen(["pidof", child_name], stdout=subprocess.PIPE, universal_newlines=True).communicate()[0].split()[0]
                 bpf_call = re.sub("{{BEFORE_PID}}", str(childpid), bpf_call)
             env = {
@@ -263,10 +264,11 @@ class Runner(object):
             #
             # Send a SIGTERM here so bpftrace exits cleanly. We'll send an SIGKILL
             # if SIGTERM didn't do the trick later
-            if p.poll() is None:
-                os.killpg(os.getpgid(p.pid), signal.SIGTERM)
-            output += p.communicate()[0]
-            result = re.search(test.expect, output)
+            if p:
+                if p.poll() is None:
+                    os.killpg(os.getpgid(p.pid), signal.SIGTERM)
+                output += p.communicate()[0]
+                result = re.search(test.expect, output)
             if not result:
                 print(fail("[  TIMEOUT ] ") + "%s.%s" % (test.suite, test.name))
                 print('\tCommand: %s' % bpf_call)
@@ -283,7 +285,7 @@ class Runner(object):
             if after and after.poll() is None:
                 os.killpg(os.getpgid(after.pid), signal.SIGKILL)
 
-        if p.returncode != 0 and not test.will_fail and not timeout:
+        if p and p.returncode != 0 and not test.will_fail and not timeout:
             print(fail("[  FAILED  ] ") + "%s.%s" % (test.suite, test.name))
             print('\tCommand: ' + bpf_call)
             print('\tUnclean exit code: ' + str(p.returncode))
